fixed the update course endpoint, added some error handling

+++ REST/DataLayer/ClientRepo.cs
@@ -44,8 +44,11 @@ namespace REST.DataLayer
         public async Task<Clients> DeleteClientById(int ClientId)
         {
             Clients client = await _context.Clients.FirstOrDefaultAsync(c => c.ClientId == ClientId);
-            _context.Clients.Remove(client);
-            await _context.SaveChangesAsync();
+            if (client != null)
+            {
+                _context.Clients.Remove(client);
+                await _context.SaveChangesAsync();
+            }
 
             return client;
         }
