+++ REST/DataLayer/OrderRepo.cs
@@ -44,9 +44,12 @@ namespace REST.DataLayer
         {
             Orders order = await _context.Orders.FirstOrDefaultAsync(o => o.OrderId == OrderId);
             OrderDetails orderDetails = _context.OrderDetails.FirstOrDefault(o => o.DetailsId == OrderId);
-            _context.OrderDetails.Remove(orderDetails);
-            _context.Orders.Remove(order);
-            await _context.SaveChangesAsync();
+            if (order != null)
+            {
+                _context.OrderDetails.Remove(orderDetails);
+                _context.Orders.Remove(order);
+                await _context.SaveChangesAsync();
+            }
 
             return order;
         }
